+++ Adapters/BookAdapter.php
@@ -5,28 +5,24 @@ class BookAdapter
 {
     public function addBook(array $arr)
     {
-        $name = $arr['name'];
+        $name = $arr['title'];
         $numPages = $arr['numPages'];
         $price = $arr['price'];
         $query = "INSERT INTO `Book` (`title`,`numPages`,`price`) VALUES ('$name','$numPages','$price')";
-        DataBase::query($query);
+        return DataBase::queryINSERT($query);
     }
 
     /**
-     * @return mixed
+     * @param $id
      */
-
-    public function getNumBook()
-    {
-        $query = "SELECT MAX(ID) FROM book";
-        $row = DataBase::query($query);
-        return $row[0];
-    }
-
     public function delBook($id)
     {
-        $query = "DELETE FROM `book` WHERE ID = ".$id;
-        DataBase::query($query);
+        $resault = false;
+        if(empty($id)) {
+            $query = "DELETE FROM `book` WHERE ID = " . $id;
+            DataBase::query($query);
+        }
+        return $resault;
     }
 
     public function editBook(array $arr,$id)
@@ -35,29 +31,68 @@ class BookAdapter
         $query = "UPDATE `book` 
                     SET ".$set." WHERE id = ".$id;
         DataBase::query($query);
-        print_r($query);
     }
-
-
     /**
      * @param $book_id
      */
-    public function delbookauthor($book_id)
+    public function delBookAuthor($book_id)
     {
         $query = "DELETE FROM `bookauthor` WHERE book_id = ".$book_id;
         DataBase::query($query);
     }
-
-
     /**
      * @param $book_id
      * @param $author_id
      */
-    public function addbookauthor($book_id, $author_id)
+    public function addBookAuthor($book_id, $author_id)
     {
         $query = "INSERT INTO `bookauthor` (`book_id`,`author_id`) VALUES 
                 ('$book_id',
                 '$author_id')";
         DataBase::query($query);
+    }
+
+    /**
+     * @param $id
+     * @return mixed
+     */
+    public function BookTitle($id)
+    {
+        $query = "SELECT title FROM book WHERE id =". $id;
+        $title = DataBase::query($query);
+        return $title[0]['title'];
+    }
+
+    /**
+     * @param $id
+     * @return int
+     */
+    public function BookNumPages($id)
+    {
+        $query = "SELECT `numPages` FROM `book` WHERE id =". $id;
+        $numPages =  DataBase::query($query);
+        return $numPages[0]['numPages'];
+    }
+
+    /**
+     * @param $id
+     * @return int
+     */
+    public function BookPrice($id)
+    {
+        $query = "SELECT `price` FROM `book` WHERE id =". $id;
+        $price = DataBase::query($query);
+        return $price[0]['price'];
+    }
+
+    /**
+     * @param $id
+     * @return mixed
+     */
+    public function BookAuthor($id)
+    {
+        $query = "SELECT `author_id` FROM `bookauthor` WHERE id =". $id;
+        $author_id = DataBase::query($query);
+        return $author_id['author'];
     }
 }
